Changed print method and expanded error messages
Changed all printf statements with a new function in the Context called
print.
Print uses by default std::cout to print all of it, you can change this by
using the setPrintMethod function and providing a pointer to which
std::ostream you want to use.

Now it is possible to directly set the output to a file, ostringstream,
etc...

With this I also expanded the error messages, giving more details on which
event caused the error and on what specific command.